feat: Move message cycling to component
